+++ core/main.ts
@@ -1,5 +1,7 @@
-export * from "./src/StateBuffer";
-export * from "./src/World";
+export * from "./src/world/DefaultWorld";
+export * from "./src/world/Game";
+export * from "./src/world/StateBuffer";
+export * from "./src/world/World";
 export * from "./src/graphics/Renderer";
 export * from "./src/graphics/Camera";
 export * from "./src/ecs/System";
