+++ core/src/contrib/entities/terrain/SpaceBackground.ts
@@ -2,8 +2,6 @@ import { Entity, Position, Renderable } from "@piggo-gg/core";
 import { Sprite, TilingSprite } from "pixi.js"
 
 export const SpaceBackground = (): Entity => {
-  const image = Sprite.from("space.png");
-  const tiles = new TilingSprite(image.texture, 6000, 6000);
 
   return {
     id: "background",
@@ -11,7 +9,12 @@ export const SpaceBackground = (): Entity => {
       position: new Position({ x: -4000, y: -2000 }),
       renderable: new Renderable({
         zIndex: -2,
-        container: async () => tiles
+        setup: async (container) => {
+          const image = Sprite.from("space.png");
+          const tiles = new TilingSprite(image.texture, 6000, 6000);
+          container.c.addChild(tiles)
+        }
+        // container: async () => tiles
       })
     }
   }
