+++ core/src/world/Game.ts
@@ -1,0 +1,9 @@
+import { Entity, System } from "@piggo-gg/core";
+
+// a game is a collection of entities and systems
+export type Game = {
+  entities: Entity[]
+  systems: System[]
+}
+
+export type GameBuilder = () => Game;
