+++ web/src/components/Header.tsx
@@ -22,8 +22,8 @@ export const Header = ({ world, netState, setNetState }: HeaderProps) => {
         Piggo
       </h1>
       <div style={{ position: 'absolute', right: 0, bottom: 0 }}>
-        <span style={{ fontFamily: "sans-serif", fontSize: 14, marginRight: 5, verticalAlign: "-80%" }}>
-          v<b>0.1.8</b>
+        <span style={{ fontFamily: "sans-serif", fontSize: 14, marginRight: 5, verticalAlign: "-70%" }}>
+          v<b>0.1.9</b>
         </span>
         <a style={{ margin: 0, color: "inherit", textDecoration: "none" }} target="_blank" href="https://discord.gg/VfFG9XqDpJ">
           <FaDiscord size={20} style={{ color: "white", verticalAlign: "-80%" }}></FaDiscord>
